metadata: change metadata a little bit
* added metadata for courses directly into the main metadata. the idea
  is to have the information (course and unit) of a file in one record
  and not spread over three records.
